Fixed Syntax Errors + Changed Return Values
Before, database functions echoed to page. Now:
For adders.php, return value is null on success and an error message on
failure;
For modifiers.php, return value is null on success and an error message
on failure;
For getters.php, return value is array(expected return value, null) on
success and array( null, error message) on failure.