getters.php, memberInfo.php/css, membersFunctions
Added a rough display logic for graceful degradation (as a backup for
the overlay - which we haven't done yet). If you click on a profile
pic, it takes you to a new page memberInfo.php which displays basic
information for that member. Literally copied most of the styling form
members.css